requirements.txt
# Core Dependencies
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Resume/Document Parsing
PyMuPDF==1.23.8
pdfplumber==0.10.0
python-docx==1.1.0
docx2txt==0.8

# NLP and Text Processing
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# AI and LLM Integration
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.7
google-generativeai==0.3.2
anthropic==0.7.8

# Vector Stores
chromadb==0.4.18
faiss-cpu==1.7.4
pinecone-client==2.2.4

# Database
sqlalchemy==2.0.23
sqlite3

# Utilities
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0